pre-commit file rename
